requirements.txt
altair==5.3.0
mysql_connector_repackaged==0.3.1
pandas==2.2.1
python-dotenv==1.0.1
Requests==2.31.0
streamlit==1.33.0
streamlit==1.32.2
